third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-inside-expected.txt

Testing parsing of the shape-inside property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("shape-inside", "none") is "none"
PASS getComputedStyleValue("shape-inside", "none") is "none"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("shape-inside", "inset(10px)") is "inset(10px 10px 10px 10px)"
PASS getComputedStyleValue("shape-inside", "inset(10px)") is "inset(10px 10px 10px 10px round 0px 0px 0px 0px / 0px 0px 0px 0px)"
PASS getCSSText("shape-inside", "inset(10px 9px)") is "inset(10px 9px 10px 9px)"
PASS getComputedStyleValue("shape-inside", "inset(10px 9px)") is "inset(10px 9px 10px 9px round 0px 0px 0px 0px / 0px 0px 0px 0px)"
PASS getCSSText("shape-inside", "inset(10px 9px 8px)") is "inset(10px 9px 8px 9px)"
PASS getComputedStyleValue("shape-inside", "inset(10px 9px 8px)") is "inset(10px 9px 8px 9px round 0px 0px 0px 0px / 0px 0px 0px 0px)"
PASS getCSSText("shape-inside", "inset(10px 20px 30px 40px)") is "inset(10px 20px 30px 40px)"
PASS getComputedStyleValue("shape-inside", "inset(10px 20px 30px 40px)") is "inset(10px 20px 30px 40px round 0px 0px 0px 0px / 0px 0px 0px 0px)"
PASS getCSSText("shape-inside", "inset(10px round 9px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 9px 9px 9px 9px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 9px 9px 9px 9px)"
PASS getCSSText("shape-inside", "inset(10px round 9px / 9px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 9px 9px 9px 9px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px / 9px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 9px 9px 9px 9px)"
PASS getCSSText("shape-inside", "inset(10px round 9px / 8px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 8px 8px 8px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px / 8px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 8px 8px 8px)"
PASS getCSSText("shape-inside", "inset(10px round 9px / 8px 7px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 8px 7px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px / 8px 7px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 8px 7px)"
PASS getCSSText("shape-inside", "inset(10px round 9px / 8px 7px 6px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 6px 7px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px / 8px 7px 6px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 6px 7px)"
PASS getCSSText("shape-inside", "inset(10px round 9px / 8px 7px 6px 5px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 6px 5px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px / 8px 7px 6px 5px)") is "inset(10px 10px 10px 10px round 9px 9px 9px 9px / 8px 7px 6px 5px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px / 7px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 7px 7px 7px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px / 7px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 7px 7px 7px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 7px 6px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 7px 6px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 7px 6px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 7px 6px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px / 7px 6px 5px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 5px 6px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px / 7px 6px 5px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 5px 6px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px / 7px 6px 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 5px 4px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px / 7px 6px 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 9px 8px / 7px 6px 5px 4px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px / 6px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 6px 6px 6px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px / 6px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 6px 6px 6px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 6px 5px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 6px 5px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 4px 5px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 4px 5px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 4px 3px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 8px / 6px 5px 4px 3px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 5px 5px 5px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 5px 5px 5px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 5px 4px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 5px 4px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 3px 4px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 3px 4px)"
PASS getCSSText("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 3px 2px)"
PASS getComputedStyleValue("shape-inside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)") is "inset(10px 10px 10px 10px round 9px 8px 7px 6px / 5px 4px 3px 2px)"
PASS getCSSText("shape-inside", "inset(10px 20px 30px 40px round 5px 6px)") is "inset(10px 20px 30px 40px round 5px 6px 5px 6px / 5px 6px 5px 6px)"
PASS getComputedStyleValue("shape-inside", "inset(10px 20px 30px 40px round 5px 6px)") is "inset(10px 20px 30px 40px round 5px 6px 5px 6px / 5px 6px 5px 6px)"
PASS getCSSText("shape-inside", "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)") is "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)"
PASS getComputedStyleValue("shape-inside", "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)") is "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)"
PASS getCSSText("shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getCSSText("shape-inside", "circle()") is "circle()"
PASS getComputedStyleValue("shape-inside", "circle()") is "circle(closest-side at 50% 50%)"
PASS getCSSText("shape-inside", "circle(farthest-side)") is "circle(farthest-side)"
PASS getComputedStyleValue("shape-inside", "circle(farthest-side)") is "circle(farthest-side at 50% 50%)"
PASS getCSSText("shape-inside", "circle(closest-side)") is "circle(closest-side)"
PASS getComputedStyleValue("shape-inside", "circle(closest-side)") is "circle(closest-side at 50% 50%)"
PASS getCSSText("shape-inside", "circle(10px)") is "circle(10px)"
PASS getComputedStyleValue("shape-inside", "circle(10px)") is "circle(10px at 50% 50%)"
PASS getCSSText("shape-inside", "circle(10px at 10px)") is "circle(10px at 10px 50%)"
PASS getComputedStyleValue("shape-inside", "circle(10px at 10px)") is "circle(10px at 10px 50%)"
PASS getCSSText("shape-inside", "circle(10px at 10px 10px)") is "circle(10px at 10px 10px)"
PASS getComputedStyleValue("shape-inside", "circle(10px at 10px 10px)") is "circle(10px at 10px 10px)"
PASS getCSSText("shape-inside", "circle(at 10px)") is "circle(at 10px 50%)"
PASS getComputedStyleValue("shape-inside", "circle(at 10px)") is "circle(closest-side at 10px 50%)"
PASS getCSSText("shape-inside", "circle(at 10px 10px)") is "circle(at 10px 10px)"
PASS getComputedStyleValue("shape-inside", "circle(at 10px 10px)") is "circle(closest-side at 10px 10px)"
PASS getCSSText("shape-inside", "circle(at top left)") is "circle(at 0% 0%)"
PASS getComputedStyleValue("shape-inside", "circle(at top left)") is "circle(closest-side at 0% 0%)"
PASS getCSSText("shape-inside", "circle(at right bottom)") is "circle(at 100% 100%)"
PASS getComputedStyleValue("shape-inside", "circle(at right bottom)") is "circle(closest-side at 100% 100%)"
PASS getCSSText("shape-inside", "circle(10px at left top 10px)") is "circle(10px at left 0% top 10px)"
PASS getComputedStyleValue("shape-inside", "circle(10px at left top 10px)") is "circle(10px at 0% 10px)"
PASS getCSSText("shape-inside", "circle(10px at top 10px left 10px)") is "circle(10px at left 10px top 10px)"
PASS getComputedStyleValue("shape-inside", "circle(10px at top 10px left 10px)") is "circle(10px at 10px 10px)"
PASS getCSSText("shape-inside", "circle(10px at right 10px bottom 10px)") is "circle(10px at right 10px bottom 10px)"
PASS getComputedStyleValue("shape-inside", "circle(10px at right 10px bottom 10px)") is "circle(10px at right 10px bottom 10px)"
PASS getCSSText("shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getCSSText("shape-inside", "ellipse()") is "ellipse()"
PASS getComputedStyleValue("shape-inside", "ellipse()") is "ellipse(closest-side closest-side at 50% 50%)"
PASS getCSSText("shape-inside", "ellipse(10px)") is "ellipse(10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px)") is "ellipse(10px closest-side at 50% 50%)"
PASS getCSSText("shape-inside", "ellipse(10px 20px)") is "ellipse(10px 20px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px)") is "ellipse(10px 20px at 50% 50%)"
PASS getCSSText("shape-inside", "ellipse(10px at 10px)") is "ellipse(10px at 10px 50%)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px at 10px)") is "ellipse(10px closest-side at 10px 50%)"
PASS getCSSText("shape-inside", "ellipse(10px 20px at 10px)") is "ellipse(10px 20px at 10px 50%)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px at 10px)") is "ellipse(10px 20px at 10px 50%)"
PASS getCSSText("shape-inside", "ellipse(10px at 10px 10px)") is "ellipse(10px at 10px 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px at 10px 10px)") is "ellipse(10px closest-side at 10px 10px)"
PASS getCSSText("shape-inside", "ellipse(at 10px)") is "ellipse(at 10px 50%)"
PASS getComputedStyleValue("shape-inside", "ellipse(at 10px)") is "ellipse(closest-side closest-side at 10px 50%)"
PASS getCSSText("shape-inside", "ellipse(at 10px 10px)") is "ellipse(at 10px 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(at 10px 10px)") is "ellipse(closest-side closest-side at 10px 10px)"
PASS getCSSText("shape-inside", "ellipse(at top left)") is "ellipse(at 0% 0%)"
PASS getComputedStyleValue("shape-inside", "ellipse(at top left)") is "ellipse(closest-side closest-side at 0% 0%)"
PASS getCSSText("shape-inside", "ellipse(at right bottom)") is "ellipse(at 100% 100%)"
PASS getComputedStyleValue("shape-inside", "ellipse(at right bottom)") is "ellipse(closest-side closest-side at 100% 100%)"
PASS getCSSText("shape-inside", "ellipse(10px at left top 10px)") is "ellipse(10px at left 0% top 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px at left top 10px)") is "ellipse(10px closest-side at 0% 10px)"
PASS getCSSText("shape-inside", "ellipse(10px at top 10px left 10px)") is "ellipse(10px at left 10px top 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px at top 10px left 10px)") is "ellipse(10px closest-side at 10px 10px)"
PASS getCSSText("shape-inside", "ellipse(10px at right 10px bottom 10px)") is "ellipse(10px at right 10px bottom 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px at right 10px bottom 10px)") is "ellipse(10px closest-side at right 10px bottom 10px)"
PASS getCSSText("shape-inside", "ellipse(10px 20px at left top 10px)") is "ellipse(10px 20px at left 0% top 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px at left top 10px)") is "ellipse(10px 20px at 0% 10px)"
PASS getCSSText("shape-inside", "ellipse(10px 20px at top 10px left 10px)") is "ellipse(10px 20px at left 10px top 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px at top 10px left 10px)") is "ellipse(10px 20px at 10px 10px)"
PASS getCSSText("shape-inside", "ellipse(10px 20px at right 10px bottom 10px)") is "ellipse(10px 20px at right 10px bottom 10px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px at right 10px bottom 10px)") is "ellipse(10px 20px at right 10px bottom 10px)"
PASS getCSSText("shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "content-box") is "content-box"
PASS getComputedStyleValue("shape-inside", "content-box") is "content-box"
PASS getCSSText("shape-inside", "padding-box") is "padding-box"
PASS getComputedStyleValue("shape-inside", "padding-box") is "padding-box"
PASS getCSSText("shape-inside", "border-box") is "border-box"
PASS getComputedStyleValue("shape-inside", "border-box") is "border-box"
PASS getCSSText("shape-inside", "margin-box") is "margin-box"
PASS getComputedStyleValue("shape-inside", "margin-box") is "margin-box"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) content-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) content-box"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) content-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) content-box"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) padding-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) padding-box"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) padding-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) padding-box"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) border-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) border-box"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) border-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) border-box"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) margin-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) margin-box"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) margin-box") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) margin-box"
PASS getCSSText("shape-inside", "content-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "content-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)"
PASS getComputedStyleValue("shape-inside", "content-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) content-box"
PASS getCSSText("shape-inside", "padding-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "padding-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)"
PASS getComputedStyleValue("shape-inside", "padding-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) padding-box"
PASS getCSSText("shape-inside", "border-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "border-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)"
PASS getComputedStyleValue("shape-inside", "border-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) border-box"
PASS getCSSText("shape-inside", "margin-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "margin-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)"
PASS getComputedStyleValue("shape-inside", "margin-box polygon(nonzero, 10px 10px, 20px 20px, 30px 30px)") is "polygon(nonzero, 10px 10px, 20px 20px, 30px 30px) margin-box"
PASS getCSSText("shape-inside", "outside-shape") is "outside-shape"
PASS getComputedStyleValue("shape-inside", "outside-shape") is "outside-shape"
PASS removeBaseURL(getCSSText("shape-inside", "url('image')")) is "url(image)"
PASS removeBaseURL(getComputedStyleValue("shape-inside", "url('image')")) is "url(image)"
PASS getCSSText("shape-inside", "calc()") is ""
PASS getComputedStyleValue("shape-inside", "calc()") is "none"
PASS getCSSText("shape-inside", "auto") is ""
PASS getComputedStyleValue("shape-inside", "auto") is "none"
PASS getCSSText("shape-inside", "rectangle()") is ""
PASS getComputedStyleValue("shape-inside", "rectangle()") is "none"
PASS getCSSText("shape-inside", "rectangle(10px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px)") is "none"
PASS getCSSText("shape-inside", "rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 10px)") is "none"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px)") is "none"
PASS getCSSText("shape-inside", "rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px 20px 30px 40px)") is "none"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "none"
PASS getCSSText("shape-inside", "inset-rectangle()") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle()") is "none"
PASS getCSSText("shape-inside", "inset-rectangle(10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px)") is "none"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 10px)") is "none"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px)") is "none"
PASS getCSSText("shape-inside", "inset-rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px 20px 30px 40px)") is "none"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "none"
PASS getCSSText("shape-inside", "inset()") is ""
PASS getComputedStyleValue("shape-inside", "inset()") is "none"
PASS getCSSText("shape-inside", "inset(10px, 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px, 10px)") is "none"
PASS getCSSText("shape-inside", "inset(10px 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px 20px, 30px)") is "none"
PASS getCSSText("shape-inside", "inset(10px, 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px, 20px 30px 40px)") is "none"
PASS getCSSText("shape-inside", "inset(10px 20px 30px 40px 50px 60px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px 20px 30px 40px 50px 60px)") is "none"
PASS getCSSText("shape-inside", "inset(round)") is ""
PASS getComputedStyleValue("shape-inside", "inset(round)") is "none"
PASS getCSSText("shape-inside", "inset(round 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(round 10px)") is "none"
PASS getCSSText("shape-inside", "inset(10px round)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px round)") is "none"
PASS getCSSText("shape-inside", "inset(10px round 10px /)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px round 10px /)") is "none"
PASS getCSSText("shape-inside", "inset(10px round 20px 30px 40px 50px 60px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px round 20px 30px 40px 50px 60px)") is "none"
PASS getCSSText("shape-inside", "inset(10px round /)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px round /)") is "none"
PASS getCSSText("shape-inside", "inset(10px round / 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(10px round / 10px)") is "none"
PASS getCSSText("shape-inside", "inset(/)") is ""
PASS getComputedStyleValue("shape-inside", "inset(/)") is "none"
PASS getCSSText("shape-inside", "inset(/ 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset(/ 10px)") is "none"
PASS getCSSText("shape-inside", "inset(round /)") is ""
PASS getComputedStyleValue("shape-inside", "inset(round /)") is "none"
PASS getCSSText("shape-inside", "circle(10px, 20px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px)") is "none"
PASS getCSSText("shape-inside", "circle(10px 20px 30px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px 20px 30px)") is "none"
PASS getCSSText("shape-inside", "circle(10px, 20px, 30px, 40px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px, 30px, 40px)") is "none"
PASS getCSSText("shape-inside", "circle(10px 20px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px 20px)") is "none"
PASS getCSSText("shape-inside", "circle(10px at 10px 10px 10px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px at 10px 10px 10px)") is "none"
PASS getCSSText("shape-inside", "circle(10px at 10px 10px at center)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px at 10px 10px at center)") is "none"
PASS getCSSText("shape-inside", "circle(10px at center center 10px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px at center center 10px)") is "none"
PASS getCSSText("shape-inside", "circle(at 10px 10px 10px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(at 10px 10px 10px)") is "none"
PASS getCSSText("shape-inside", "circle(at 10px 10px at center)") is ""
PASS getComputedStyleValue("shape-inside", "circle(at 10px 10px at center)") is "none"
PASS getCSSText("shape-inside", "circle(at center center 10px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(at center center 10px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px, 20px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px, 30px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px 30px 40px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px 20px 30px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px 30px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px at 10px 10px 10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px at 10px 10px 10px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px at 10px 10px at center)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px at 10px 10px at center)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px at center center 10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px at center center 10px)") is "none"
PASS getCSSText("shape-inside", "ellipse(10px 20px 30px at center center 10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px 30px at center center 10px)") is "none"
PASS getCSSText("shape-inside", "ellipse(at 10px 10px 10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(at 10px 10px 10px)") is "none"
PASS getCSSText("shape-inside", "ellipse(at 10px 10px at center)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(at 10px 10px at center)") is "none"
PASS getCSSText("shape-inside", "ellipse(at center center 10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(at center center 10px)") is "none"
PASS getCSSText("shape-inside", "polygon()") is ""
PASS getComputedStyleValue("shape-inside", "polygon()") is "none"
PASS getCSSText("shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is "none"
PASS getCSSText("shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is "none"
PASS getCSSText("shape-inside", "polygon(nonzero)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero)") is "none"
PASS getCSSText("shape-inside", "polygon(evenodd)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd)") is "none"
PASS getCSSText("shape-inside", "polygon(10px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(10px)") is "none"
PASS getCSSText("shape-inside", "polygon(nonzero,10px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero,10px)") is "none"
PASS getCSSText("shape-inside", "polygon(evenodd,12px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd,12px)") is "none"
PASS getCSSText("shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "none"
PASS getCSSText("shape-inside", "content-box content-box") is ""
PASS getComputedStyleValue("shape-inside", "content-box content-box") is "none"
PASS getCSSText("shape-inside", "polygon(nonzero, 0 0 ,10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 0 0 ,10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "none"
PASS getCSSText("shape-inside", "none content-box") is ""
PASS getComputedStyleValue("shape-inside", "none content-box") is "none"
PASS getCSSText("shape-inside", "content-box none") is ""
PASS getComputedStyleValue("shape-inside", "content-box none") is "none"
PASS getCSSText("shape-inside", "polygon(none) content-box") is ""
PASS getComputedStyleValue("shape-inside", "polygon(none) content-box") is "none"
PASS getCSSText("shape-inside", "content-box polygon(none)") is ""
PASS getComputedStyleValue("shape-inside", "content-box polygon(none)") is "none"
PASS getCSSText("shape-inside", "circle(50px) none") is ""
PASS getComputedStyleValue("shape-inside", "circle(50px) none") is "none"
PASS getCSSText("shape-inside", "none circle(50px)") is ""
PASS getComputedStyleValue("shape-inside", "none circle(50px)") is "none"
PASS getCSSText("shape-inside", "url('shape.svg') content-box") is ""
PASS getComputedStyleValue("shape-inside", "url('shape.svg') content-box") is "none"
PASS getCSSText("shape-inside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("shape-inside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "none"
PASS getParentAndChildComputedStylesString("shape-inside", "none", "rectangle(10px, 20px, 30px, 40px)") is "parent: none, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "outside-shape", "rectangle(10px, 20px, 30px, 40px)") is "parent: outside-shape, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "initial") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: none"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: none"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "inherit") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "", "inherit") is "parent: none, child: none"
PASS getParentAndChildComputedStylesString("shape-inside", "none", "inherit") is "parent: none, child: none"
PASS successfullyParsed is true

TEST COMPLETE
